using external image ratio calculator
